requirements.txt
# ===== Core Framework =====
Flask==3.1.0
Werkzeug==3.1.3
flask-cors==5.0.1

# ===== Authentication & Security =====
PyJWT==2.9.0
bcrypt==4.2.1
Flask-WTF==1.2.2
flask-talisman==1.1.0
flask-limiter==3.8.0
bleach==6.2.0

# ===== Input Validation =====
marshmallow==3.23.2

# ===== Task Queue =====
Celery==5.4.0
redis==5.2.1

# ===== Database =====
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
pymysql==1.1.1
mysql-connector-python==9.1.0
alembic==1.14.1

# ===== Data & ML =====
Pandas==2.2.3
numpy==1.26.4
torch==2.1.0
transformers==4.46.3
datasets==3.1.0
safetensors==0.4.5

# ===== Scraping & HTTP =====
requests==2.32.3
beautifulsoup4==4.12.3
aiohttp==3.11.10

# ===== Search & Monitoring =====
elasticsearch==8.16.0
sentry-sdk==2.21.0

# ===== Encryption =====
cryptography==44.0.0

# ===== Utilities =====
python-dotenv==1.0.1
pillow==11.0.0
pyTelegramBotAPI==4.24.0
